Split the data operation and save-drawing suites out of NoteStoreTests
